fixup! client: abstract generation
